+++ src/main/java/com/github/kory33/chatgui/exception/ChatInterceptionCancelledException.java
@@ -1,4 +1,4 @@
-package com.github.kory33.signvote.exception;
+package com.github.kory33.chatgui.exception;
 
 import lombok.Getter;
