+++ src/main/java/com/github/kory33/chatgui/model/player/BrowseablePageInterface.java
@@ -1,4 +1,4 @@
-package com.github.kory33.signvote.ui.player.model;
+package com.github.kory33.chatgui.model.player;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -6,9 +6,9 @@ import java.util.List;
 import lombok.Setter;
 import org.bukkit.entity.Player;
 
-import com.github.kory33.messaging.tellraw.MessagePartsList;
-import com.github.kory33.signvote.collection.RunnableInvoker;
-import com.github.kory33.signvote.manager.PlayerInteractiveInterfaceManager;
+import com.github.kory33.chatgui.tellraw.MessagePartsList;
+import com.github.kory33.chatgui.command.RunnableInvoker;
+import com.github.kory33.chatgui.manager.PlayerInteractiveInterfaceManager;
 import com.github.ucchyocean.messaging.tellraw.MessageParts;
 
 import lombok.Getter;
