+++ src/main/java/com/github/kory33/signvote/command/subcommand/AddScoreCommandExecutor.java
@@ -2,7 +2,7 @@ package com.github.kory33.signvote.command.subcommand;
 
 import java.util.ArrayList;
 
-import com.github.kory33.signvote.collection.RunnableInvoker;
+import com.github.kory33.chatgui.command.RunnableInvoker;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
@@ -12,12 +12,12 @@ import com.github.kory33.signvote.constants.MagicNumbers;
 import com.github.kory33.signvote.constants.MessageConfigNodes;
 import com.github.kory33.signvote.constants.PermissionNodes;
 import com.github.kory33.signvote.core.SignVote;
-import com.github.kory33.signvote.listeners.PlayerChatInterceptor;
-import com.github.kory33.signvote.manager.PlayerInteractiveInterfaceManager;
+import com.github.kory33.chatgui.listener.PlayerChatInterceptor;
+import com.github.kory33.chatgui.manager.PlayerInteractiveInterfaceManager;
 import com.github.kory33.signvote.manager.VoteSessionManager;
 import com.github.kory33.signvote.session.VoteSession;
 import com.github.kory33.signvote.ui.player.AddScoreInterface;
-import com.github.kory33.signvote.ui.player.model.FormChatInterface;
+import com.github.kory33.chatgui.model.player.FormChatInterface;
 
 /**
  * Executor class of "addscore" sub-command
