+++ src/main/java/com/github/kory33/signvote/command/subcommand/StatsCommandExecutor.java
@@ -2,8 +2,7 @@ package com.github.kory33.signvote.command.subcommand;
 
 import java.util.ArrayList;
 
-import com.github.kory33.signvote.collection.RunnableInvoker;
-import com.github.kory33.signvote.collection.RunnableInvoker;
+import com.github.kory33.chatgui.command.RunnableInvoker;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
@@ -13,10 +12,10 @@ import com.github.kory33.signvote.constants.MessageConfigNodes;
 import com.github.kory33.signvote.constants.PermissionNodes;
 import com.github.kory33.signvote.constants.StatsType;
 import com.github.kory33.signvote.core.SignVote;
-import com.github.kory33.signvote.manager.PlayerInteractiveInterfaceManager;
+import com.github.kory33.chatgui.manager.PlayerInteractiveInterfaceManager;
 import com.github.kory33.signvote.manager.VoteSessionManager;
 import com.github.kory33.signvote.session.VoteSession;
-import com.github.kory33.signvote.ui.player.model.PlayerClickableChatInterface;
+import com.github.kory33.chatgui.model.player.PlayerClickableChatInterface;
 import com.github.kory33.signvote.ui.player.stats.StatsInterface;
 
 /**
