+++ src/main/java/com/github/kory33/signvote/manager/VoteSessionManager.java
@@ -9,7 +9,7 @@ import java.util.stream.Stream;
 
 import org.bukkit.block.Sign;
 
-import com.github.kory33.signvote.collection.BijectiveHashMap;
+import com.github.kory33.chatgui.util.collection.BijectiveHashMap;
 import com.github.kory33.signvote.model.VotePoint;
 import com.github.kory33.signvote.session.VoteSession;
 import com.github.kory33.signvote.utils.FileUtils;
