+++ src/main/java/com/github/kory33/signvote/ui/player/VoteInterface.java
@@ -8,8 +8,8 @@ import com.github.kory33.signvote.ui.player.defaults.IDefaultClickableInterface;
 import lombok.Getter;
 import org.bukkit.entity.Player;
 
-import com.github.kory33.messaging.tellraw.MessagePartsList;
-import com.github.kory33.signvote.collection.RunnableInvoker;
+import com.github.kory33.chatgui.tellraw.MessagePartsList;
+import com.github.kory33.chatgui.command.RunnableInvoker;
 import com.github.kory33.signvote.configurable.JSONConfiguration;
 import com.github.kory33.signvote.constants.MessageConfigNodes;
 import com.github.kory33.signvote.exception.InvalidScoreVotedException;
@@ -18,7 +18,7 @@ import com.github.kory33.signvote.exception.VotePointAlreadyVotedException;
 import com.github.kory33.signvote.exception.VoteSessionClosedException;
 import com.github.kory33.signvote.model.VotePoint;
 import com.github.kory33.signvote.session.VoteSession;
-import com.github.kory33.signvote.ui.player.model.PlayerClickableChatInterface;
+import com.github.kory33.chatgui.model.player.PlayerClickableChatInterface;
 import com.github.ucchyocean.messaging.tellraw.MessageParts;
 
 /**
